Rework artifact inputs
Gradle's `@Classpath` annotation seems quite strange and turns out is not actually what we want. This reworks things to explicitly use `ArtifactCollection` + absolute paths for caching. This is important because we care about the jar locations even if the contents are the same because we need to read their contents during task exec. This is a nicer solution than just disabling caching entirely as it will still allow caching on the same machine.